+++ src/components/domains/myPage/layout/SidebarModal.module.scss
@@ -1,0 +1,3 @@
+.sidebar-modal {
+  height: 100%;
+}
