+++ src/layout/MyPageSubLayout/MyPageSubLayout.module.scss
@@ -37,10 +37,6 @@
     @include responsive("T") {
       display: none;
     }
-
-    &--mobile {
-      height: 100%;
-    }
   }
 
   &__main {
